restructure tab to ipo and media block
