--- scripts/mock-recent-ipodata.json
@@ -1,38 +1,0 @@
-[
-    {
-        "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
-        "IpoFullName": "Bharti Recent Hexacom IPO-1",
-        "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
-    },
-    {
-        "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
-        "IpoFullName": "Bharti Recent Hexacom IPO-2",
-        "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
-    },
-    {
-        "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
-        "IpoFullName": "Bharti Recent Hexacom IPO-3",
-        "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
-    },
-    {
-        "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
-        "IpoFullName": "Bharti Recent Hexacom IPO-4",
-        "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
-    },
-    {
-        "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
-        "IpoFullName": "Bharti Recent Hexacom IPO-5",
-        "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
-    }
-  ]
-
