+++ scripts/mock-upcoming-ipodata.json
@@ -3,36 +3,24 @@
         "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
         "IpoFullName": "Bharti Upcoming Hexacom IPO-1",
         "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
+        "IPOEventEndDate": "05 Apr 2024"
     },
     {
         "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
         "IpoFullName": "Bharti Upcoming Hexacom IPO-2",
         "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
+        "IPOEventEndDate": "05 Apr 2024"
     },
     {
         "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
         "IpoFullName": "Bharti Upcoming Hexacom IPO-3",
         "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
+        "IPOEventEndDate": "05 Apr 2024"
     },
     {
         "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
         "IpoFullName": "Bharti Upcoming Hexacom IPO-4",
         "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
-    },
-    {
-        "IPOLogoImage": "https://www.icicidirect.com/images//Bharti Hexacom Logo-202403271411036233854.png",
-        "IpoFullName": "Bharti Upcoming Hexacom IPO-5",
-        "IPOEventStartDate": "03 Apr 2024",
-        "IPOEventEndDate": "05 Apr 2024",
-        "KnowMore": "https://www.icicidirect.com/ipo"
+        "IPOEventEndDate": "05 Apr 2024"
     }
   ]
-
